Set up seed data & deploying to Heroku (#6)
* Set up seed data & deploying to Heroku

https://eaflood.atlassian.net/browse/WC-121

The app relies on both registrations and users being created by the existing service. We'll need some seed data for existing users and their registrations so we can test the app independently. This change covers any config changes required to deploy to Heroku for testing.

Several field names didn't match what was currently in the database, so these had to be updated in models and tests. We should also try to alias CamelCase field names from the DB to snake_case when possible.