Add transient object for renewals (#7)
https://eaflood.atlassian.net/browse/WC-125

We want to store information about a renewal as a user goes through the process, but we don't want to modify the existing registration until that process is complete. So we need an additional object to be stored in the database to capture this incomplete renewal.

This change also creates a starting page for the renewals process. If the reg_identifier provided matches a registration which can be renewed, clicking "Continue" will create a new TransientRegistration to store this renewal information.