Set up 'happy path' for limited company renewals (#8)
https://eaflood.atlassian.net/browse/WC-124

The 'happy path' should cover the pages which the user will need to complete a renewal, provided they meet certain conditions (in this case, a limited company based in the UK). This is in a fairly skeletal form - just a heading, a back link and a continue button. It should provide a basis for testing and further development. Functionality like the address lookup and smart answers is not covered.

Users should be able to move through the forms in the correct order. This is done by implementing a state machine to manage how the user travels between pages.

Most of the forms and their assorted files were generated using a custom-written scaffolding script, which is stored in the waste carriers tools repo.