Renewal start page and user permissions (#11)
https://eaflood.atlassian.net/browse/WC-89

Before users begin the renewal process, we should set their expectations for how long the renewal lasts and what it will cost.

We should also make sure they actually have permission to renew the registration – users should not be able to modify registrations which don't belong to them. I've implemented CanCanCan to handle this, which might be overkill for now but will come in handy once we have to handle admin users as well.

This change also adds some custom error pages as there were several instances where it didn't make sense to display the errors on the page (as with validations).