Improve database seeding (#37)
https://eaflood.atlassian.net/browse/WC-167

We need some seed data for the database for the dev environment and for running acceptance tests against. This change switches from loading all seeds in a single script to storing seeds as multiple JSON files and then loading them. This approach should make it easier to add and modify seeds.

Seeds with regIdentifiers are loaded first. Then ones without regIdentifiers are loaded and assigned new regIdentifiers to 'fill in the gaps'.

This change also includes a refactor of the CanGenerateRegIdentifier concern to continue incrementing the counter if it would clash with a regIdentifier which was added out of sequence.