Implement smart answers (#38)
https://eaflood.atlassian.net/browse/WC-92

This change adds the "smart answers" forms to help determine if a user is an upper tier or lower tier waste carrier.

This also includes a refactor of form objects. Form model subclasses should now be much slimmer as they only need to deal with attributes which are specific to that form, while the BaseForm deals with updating the transient_registration and with common attributes like the reg_identifier.

Also, validation errors should now meet the standards in http://govuk-elements.herokuapp.com/errors/ This included a fix for the business type form as well as this had been previously implemented.