Refactor model tests for Transient Registration (#44)
This file has become bloated and repetitive, mostly due to the bulk of the state change tests. This change refactors it to remove duplication and split it into separate files when necessary.

The business_type_form state tests were especially repetitive as we're just feeding in different values and making sure we get the correct result. I switched to using a permutation table pattern as described in this talk: https://www.youtube.com/watch?v=KjENZNjRCWM This allows us to just store the values and the expected result in a single hash, and then run them all through the same process.