Get the registration number of a limited company or LLP (#49)
https://eaflood.atlassian.net/browse/WC-97

Limited companies and LLPs must provide their registration number for Companies House. We then need to validate that the number is in the correct format, then query Companies House to make sure the company/LLP is active. We should then inform the user if Companies House says there's a problem, or continue to the next stage of the form if all is well.