Changing company's registration number exits renewals process (#57)
https://eaflood.atlassian.net/browse/WC-98

If the company's registration number changes, it is considered a new entity. This means the registration cannot be renewed and they must make a new one instead.

It was previously valid to have company numbers with fewer than 8 digits – we need to account for this when we check if the number has changed (eg 123456 should be considered the same as 00123456).