Search for address by postcode (#62)
https://eaflood.atlassian.net/browse/WC-100

UK-based users should be able to get their address by searching with the postcode. This saves time typing in the address, but also ensures data that is more compliant with our standards.

The AddressFinderService sends a request to an instance of our OS Places address lookup wrapper (https://github.com/DEFRA/os-places-address-lookup). The location of this should be set in the .env file.

If the user selects an address from the dropdown and saves it, this address should be pre-selected when they return to the page. Same goes for the postcode used to make the lookup.