Fix content on login page (#76)
https://eaflood.atlassian.net/browse/WC-201

The content on the login page should match the approved content on the wireframe. This includes information about confirming your email address. User accounts are created and confirmed in waste-carriers-frontend. We don't want to duplicate all the config for configuration emails in the renewals app at this time, so we link to the frontend app instead.