Provide an address manually (#74)
https://eaflood.atlassian.net/browse/WC-101

While we prefer users to add addresses using the OS Places API, some users will not be able to use it, for example if they are overseas, if their address is not listed or if the service is down. These users should be able to fill in a separate form with details of their address.

*Pre-filling forms*

Non-overseas users can only modify the postcode on the postcode form. Otherwise we'd be asking them for the postcode twice. Instead we just show the postcode they've already set on the page.

If the user provides an address, saves and then goes back to the form later, we want to fill in this address. However, if the user has since changed the postcode on the postcode form, we don't want to fill in the old address as this will have changed. Instead we should just fill in the postcode which they provided. This does not apply to overseas users, who would never be setting a postcode on a different form.

*Routing*

Users with the 'overseas' business type will get the manual address form by default.

If the OS Places lookup returns addresses but the user doesn't want to use any of them, they can choose to go to the manual address form instead.

If the postcode format is fine but no results are found, users also get the option to enter the address manually. A postcode search with no results is technically invalid – however, we save the entered postcode anyway in case the user needs it when entering the postcode manually.

If a user searches with a postcode and we get an error back, we redirect to the manual form instead. This is done by temporarily persisting a flag for the error against the transient_object and use this to route the user to the manual address form. This flag is then removed once the manual address form is loaded. Not ideal, but also means less mucking around in the controller, and the way transitions are handled is much more consistent.