Change 'other' business type to 'charity or trust' (#80)
https://eaflood.atlassian.net/browse/WC-202

Charities and trusts are the only ones who should select this option. We should rename this business type so it's clear who this option applies to, rather than it being a catchall.

Renaming 'other' to 'charity' also matches our existing data, so we can prefill the form for the user and don't need extra logic to determine valid changes.