Generate relative dates for seed files (#88)
https://eaflood.atlassian.net/browse/WC-214

When seeding registrations, we sometimes want their dates to be relative to the current date (eg 3 months ago) as opposed to a specific date (eg 26 February 2018). This change lets us specify relative dates in seed data and parse the seed data appropriately.

When we parse each JSON seed, check to see if the "date_flag" is set and then generate all time-sensitive dates based on the flag. Options are expired, in the renewal window or outside the renewal window.