Add multiple key people (#93)
https://eaflood.atlassian.net/browse/WC-106

Most registrations (excluding sole traders) should be able to add multiple key people. Partnerships must have a minimum of 2.

This doesn't cover removing key people, just adding multiple ones.

This change adds a new controller action and route which submits the form, but then reloads it, allowing the user to add multiple people. Anyone new is appended to the list of existing people, unless validations fail.

*Buttons*

The "Add another" button triggers the new controller action. This button is not visible if only one key person can be added (like sole traders).

Sometimes "Add another" will be the main action (for example, when adding the first partner) but most of the time it's a secondary action. The styling of this button changes depending on this.

If the user must add more people, the "Continue" button isn't visible yet.

*Listing added people*

If the business type allows multiple key people and at least one person has been added, these are listed under the "You have added the following people" heading.

If only one key person can be added at any one time, saving that person and then returning to the form should pre-fill the fields with the existing details.

*Validating key people*

Partnerships must have at least 2 key people. All others must have at least 1. We have validation to make sure enough people have been added before proceeding to the next stage.

If the user attempts to add a completely blank person, we check to see if the minimum number of people has already been reached. If it is, the form validates but does not add the blank person. If it does not validate, the user is prompted to add another person.

*Other changes*

Also reworked how the date fields are processed to make sure we set nil values if the date fields aren't in the correct format - otherwise it didn't overwrite any existing values (eg if a sole trader saved a key person, then tried to edit it).

Finally, updated VCR cassettes.