Check which country users are based in (#97)
* Check which country users are based in

https://eaflood.atlassian.net/browse/WC-238

This PR adds a new form to ask where the business is based - England, Wales, Scotland, Northern Ireland or overseas.

Wales, Scotland and Northern Ireland have their own registers for waste carriers. This form will allow us to direct those users to the correct service which they should be registering with. Users can continue to register with the EA if they want, but this will at least make them aware of the other registry.

Overseas users operating in England should register with the EA - however, there are some differences in what forms they see (for example, the address forms). Currently we filter them out using the 'overseas' option on the business type form, but 'overseas' makes a lot more sense as a location than a business type.

* Update VCR cassettes

* Scaffold location form

Set up the required fields for the form and modify routing so it fits between the renewal start page and the business type form.

* Set up form fields and save data

* Validate that a valid location has been selected

* Set up Wales, Scotland and NI pages

Set up all the required files and tests for these pages. We still need to add content, and there's also no way to access these pages yet - you can get out, but not in.

* Route users from location form to country pages

Users who say they are based in NI, Scotland or Wales are sent to the relevant page for that country.

Users from England are sent to the business type form.

Users from overseas skip the business type form and go straight to smart answers.

* Route users back to location pages

Users hitting 'back' should see the same pages that they saw when hitting 'forward'. This change means that any user who was sent through a particular "You should register in X" page, or skipped the business type form because they were overseas, should get the correct user journey in reverse too.

* Remove checks for 'overseas' business type

Since we are now flagging overseas users based on the location, we should check that field instead of business_type to filter for overseas users. This commit changes a lot of logic which was relying on checking the business_type and instead introduces an `overseas?` method which relies on the location instead.

We do continue to set the business_type to overseas as this is used for microcopy in our app, and is expected to be filled in when other applications access registrations from the shared database. However, we now set this in the location form if the location is overseas.

* Add content to other location pages

This is taken from our wireframe and should be reviewed by a content designer once we have one.