Declare if a relevant environmental offence exists (#100)
https://eaflood.atlassian.net/browse/WC-107

"Has anyone involved in managing your organisation been convicted of an environmental offence in the past 12 months?" Users can only answer yes/no on this form – if they say yes, they should be sent to an additional form to provide further details (to be dealt with in a later story). If a user has no convictions to declare, they can skip the conviction details form and go straight to the contact name form.

Any conviction information is only relating to the last 12 months from the date we ask/check. This information will be 3 years out of date when the user renews, so we have no reason to import it. We should ask for updated information instead.