Enter the name of the person to contact (#110)
https://eaflood.atlassian.net/browse/WC-109

"Who should we contact about this registration?" Allow users to provide the name of a contact person for their registration and renewal. Fields should be pre-populated with the name the user previously entered. Includes validations to make sure values are present, are not too long and only contain valid characters.