Rename 'key people' (#111)
Currently we have KeyPeople, which is all people. But we also have 'key people', which are KeyPeople with the person_type of 'key', and "people with relevant convictions", which are KeyPeople with the person_type of 'relevant'.

This is confusing so we've decided to rename 'key people' to 'main people'. The person_type of 'key' will have to stay so we can remain compatible with other applications using the DB. However, the form object and other various methods should be renamed for clarity's sake.