Clean data when creating and saving transient registrations (#112)
* Clean data when creating and saving transient registrations

This change adds some features to improve our data quality.

This means removing excess whitespace at the start or end of the field. Some forms do this on saving but others don't, so we should try to implement this across all forms in a standardised way instead. We should also do this when we import data from a Registration to a TransientRegistration object.

There are also some fields, like phone numbers, which previously have not been validated as closely as we would have liked. It doesn't make sense to prefill a field with a value which will then be considered invalid, so we should run some validations on importing the data when the TransientRegistration is created.

* Strip whitespace on submitting a form

* Refactor to avoid unnecessary 'if' calls

* Strip whitespace from attributes when importing transient registration

If the attributes we're importing from a registration have excess whitespace at the start or end, clean them up before importing them.

Decided to move the whitespace stripping methods to a concern as they're being shared by the form objects and the TransientRegistration model.

Also removed unnecessary helper file.

* Remove invalid attributes from brand new transient registrations

When we create a new transient registration, we import attributes from the source registration. Due to lax validations in the past, there's a possibility that some of this data may not pass our new validations. Prefilling a field with data which does not validate is frustrating for the user so we've chosen to remove it instead. For more info on this decision, see https://eaflood.atlassian.net/browse/WC-131

Phone numbers were the main reason for this, so I've also implemented phone_number validation so we have something to test with. I used the Phonelib gem for consistency with our other Rails apps (WEX and FRAE).