Add contact email address (#116)
* Add contact email address

https://eaflood.atlassian.net/browse/WC-111

Implement form for "What's the email address of the person we should contact?". This form should be prefilled with the contact_email which was originally provided. The field should also validate as being in the correct format for an email address, and be tagged as an "email" field so mobile users get the correct keyboard.

* Set up and persist form

We want users to confirm their email address, so we have an additional field for this. The field is also prepopulated with the existing answer, so users who aren't changing the email won't have to type anything but can just click "continue".

* Validate email fields

First we check that the fields are filled in. Then we check if the email format is correct, and if the two fields match.

We only check the format of the contact_email field, not the confirmation field, as that only really exists to make sure they match. If contact_email is a correct format and confirmed_email is identical to that, then it would also be a correct format.

The email validation gem also covers length, so we don't need an additional check for that.