Search for a contact address by UK postcode (#117)
* Search for a contact address by UK postcode

https://eaflood.atlassian.net/browse/WC-112

"What's the address of the person we should contact?" Add forms to allow UK-based users to enter a postcode so they can search for a contact address. Selection and saving of the contact address will be added in a separate PR to keep things manageable.

This should function the same way as the registered address forms earlier in the process.

Non-UK users or users who choose to enter their addresses manually will see a separate form (to be worked on as a separate feature).

* Set up contact postcode form

Currently this is just a copy of the registered address postcode form with names swapped around (and some tweaks to the postcode validator to deal with the two different postcode fields).

References to the manual contact address page have been commented out until that page is created.

Next step is to refactor to remove masses of duplication.

* Rename temp_postcode to temp_company_postcode

This is the attribute used for looking up registered addresses. Now that we have two temporary postcodes, I've renamed this to be a bit clearer.

* Rename and refactor postcode validator

Make it less specific to certain attributes, and also reduce complexity of the validate method.

* Add explanatory text to page

* Postcode forms now inherit from PostcodeForm