Enter contact address manually (#118)
* Enter contact address manually

https://eaflood.atlassian.net/browse/WC-113

Overseas users, or users who can't use the postcode lookup for whatever reason, should have the opportunity to enter their address manually.

This should behave the same way as the manual address entry form for the registered address.

* Set up contact_address_manual form

Currently this is a copy of the registered address manual forms, but renamed as appropriate.

References to the contact postcode form are commented out until we merge https://github.com/DEFRA/waste-carriers-renewals/pull/117

* Manual address forms now inherit from ManualAddressForm

Create a subclass of BaseForm which CompanyAddressManualForm and ContactAddressManualForm inherit from, to reduce dupliation as both forms are very similar. (Turns out there was a lot of duplication!)

* Move fields for manual address entry into shared view partial

* Allow users to skip from contact postcode to manual address form

If OS Places doesn't have any results for the postcode search, include a link to let users enter the address themselves.