Add contact telephone number (#115)
https://eaflood.atlassian.net/browse/WC-110

This change adds the form for "What's the telephone number of the person we should contact?"

We also use a `telephone_field` instead of `text_field`. This adds a `type='tel'` tag to the rendered HTML, which should mean that mobile users get an appropriate keyboard when the field is selected.

We use Phonelib to validate phone numbers to make sure they exist, are not too long, and are also in a valid format.