Select contact address from postcode lookup results (#120)
https://eaflood.atlassian.net/browse/WC-112

Once users have entered a postcode to search with, we should let them choose the address they want from the results and save that as their contact address.

This should work the same way as the registered address earlier in the form.

We should go with the same inheritance model we've used for the people forms, postcode forms and manual address forms. CompanyAddressForm and ContactAddressForm are very similar, so any shared methods should be inherited from the AddressForm instead.