Set up 'check your details' page (#121)
* Set up 'check your details' page

https://eaflood.atlassian.net/browse/WC-114

Once users have finished filling in their details, they should be given the opportunity to check and change their details before paying and submitting.

* First pass at displaying all answers

This is the first attempt to display all the details which the user has entered or confirmed. We don't display smart answers as it can be difficult to detect which answers influenced the final decision if the user hits 'back' and goes down a different path.

Needs designing. Badly.

* First pass at styling the 'check your details' page

Split the information into sections to match the wireframe. I added some missing info and also removed the smart answers results, as it's difficult to tell what answers were actually used if the user goes back and changes their responses.

* Display error if renewal is missing some information

Users shouldn't be able to get this far without answering all the required questions. However, if something has gone wrong, this error lists the missing fields and asks them to go back and fix it.