Add declaration page (#123)
https://eaflood.atlassian.net/browse/WC-115

Before submitting, users must agree to a declaration that the information they've submitted is accurate and they understand the terms. This is done with a checkbox, and we validate that the user has checked it. We don't import their previous answer as they need to answer the declaration for this renewal.