Implement version page (#127)
* Implement version page

https://eaflood.atlassian.net/browse/WC-264

This page is a standard feature in our Rails apps that displays the version and commit hash of the currently-running code. This is a diagnostic tool to confirm we are running the correct version, but is also used for external monitoring as a simple, low-impact way of confirming the status of the app.

* Set up skeleton of page

* Add application version and current commit

This will need further work when we move the code into an engine - we'll need to keep track of the engine version separately from the frontend and backend versions. But for now it's simple enough to just pop in the config.

The current_git_commit code is cribbed from waste-exemptions-shared.

* Store application name in config

* Tests for ApplicationHelper, and fix survey link

Added some tests for the ApplicationHelper. These will need improving but it's a start!

Also fixed the link for the survey while I was in there as it was set to go to the wrong project.

* Move page to /pages/version