Checks for can_renew include expires_on (#132)
https://eaflood.atlassian.net/browse/WC-185

Currently there will be a brief window when a registration is officially EXPIRED but it is still recorded as ACTIVE in the service. This is due to the timing of when the job that marks expired registrations is run.

To avoid any issues with someone renewing before the expiry job runs, we should make sure that a registration cannot be renewed if the current date is equal to or after the expires_on date.

When we check if a registration should be expired, we want to compare against the current date in the UK. This may be different than the current date in UTC when daylight savings is in effect.

If a user registers around the time of year when DST usually starts or ends, there's also a chance that they may be registering in one time zone and expiring in another. So we need to account for this to make sure registrations don't gain or lose a day.

This PR also stops expired registrations from being renewed, as this had previously been an option.