Make smart answers optional (#136)
* Make smart answers optional

https://eaflood.atlassian.net/browse/WC-275

In order to reduce the number of pages users have to go through, and to give more choice for those who have good knowledge of what the difference is between lower and upper tier, we will make smart answers optional.

We have already tested this in user research without issue, and the team feels it better highlights that smart answers is an aid and not an approval or clarification of what users can do (which testing has shown some view it as).

This story covers implementing the version contained in the prototype into the production renewals service.

* Scaffold the new tier_check_form

* Set up and persist the tier check form

Add text and form fields, and store the user's answer so we can prefill the form if they return to it.

The tier check form uses a new temporary attribute called temp_tier_check for the answer - this will be used for routing into or around smart answers in the future. Adding that routing is the next step.

Also update the .gitignore to get rid of pesky .DS_Store files.

* Add missing error message

* Route around smart answers based on tier check answer

If a user asks to skip smart answers, route them straight to the CBD type page instead.

If a user tries to go back from the CBD page, cehck if they have skipped smart answers and send them straight back to the tier check page if so. Otherwise, send them back through the smart answers route they came from, just as we did before.