Update project to support connecting to MongoDb 3.6 (#141)
* Correct mongoid.yml config file format

Since we first built the renewals service we have been unable to connect to MongoDb with authentication switched on.

Having investigated the issue, it all appears to be due to changes in the format of the `mongoid.yml` file caused by Mongoid being taken in house (by MongoDB), and replacing the underlying driver Moped with the official ruby MongoDb one.

These changes resolve those formatting issues, and tests confirm that it can now connect to a MongoDb 3.6 database with authentication enabled.

The changes also cover removing the reference to mongo in the Gemfile. Mongoid relies on the mongo gem as its driver for connecting to MongoDb, and as such it will bring in the mongo gem (and actually later version than 2.4.2).

As such there is no need to specify it in the Gemfile so this change removes it and updates the project.