Add payment summary page (#145)
* Add payment summary page

https://eaflood.atlassian.net/browse/WC-116

Before they are asked for payment, users should see a summary of the amount they must pay and their options for paying it.

This should include the renewal fee, the cost of any copy cards, and the fee for changing the registration type (if applied).

This does not cover setting up the order or payment objects in the database as this will be dealt with when WorldPay is implemented.

* Set up form values and persist them

A first pass at setting up the content for the page and saving the user's answer about how they'd like to pay.

The copy cards fee depends on a different PR being merged in (https://github.com/DEFRA/waste-carriers-renewals/pull/139). For now we've got a temporary holding value.

* Content changes to payment summary and WorldPay substitute

Confirm how many cards the user has ordered on the payment summary page.

Also add some explanatory copy to the WorldPay substitute page explaining what would happen here once WorldPay is implemented. This is temporary but should make it a bit clearer what is going on during testing.

* Move card charge to env vars

Move the charge for registration cards to env vars, with the rest of the cards.

Also make some minor tweaks now that the registration card page has been merged in.