Pay for a renewal by bank transfer (#146)
* Pay for a renewal by bank transfer

https://eaflood.atlassian.net/browse/WC-118

If a user chooses to pay by bank transfer on the Payment Summary page, they should be sent to an additional page which has all the necessary instructions, information and bank details for them to make the payment.

This is instead of going through WorldPay.

It should be made clear that paying by bank transfer means you cannot instantly renew your registration as you have to wait for your renewal to be processed.

* Scaffold bank transfer form

* Route users to bank transfer based on payment method

* Add page content

Add all the information about how to pay by bank transfer. The only thing missing is the total cost - we still need to calculate this and add it.

* Calculate total cost to be paid

This had already been done in a separate branch: https://github.com/DEFRA/waste-carriers-renewals/pull/145/ but it made more sense to move it from the form into the transient_registration.

This also involved renaming some existing methods in the transient_registration for clarity's sake.

* Minor tweaks to make page clearer

Split addresses onto multiple lines so they are easier to read and the table is less squashed.

Also add a line to clarify that the renewal isn't complete until both payment steps are done.

* Refactor to include changes from payment_summary_form

Now that the payment summary form branch is merged in, we can update our routing tests.

Also make some tweaks to how the charge information is stored.