'Application received' page for renewals with a pending conviction check or payment (#149)
* Complete a renewal with a pending conviction check or payment

https://eaflood.atlassian.net/browse/WC-120

If a user submits their renewal and pays by bank transfer, or submits and has a conviction check pending, they should see a "Renewal received" page.

This story covers setting up that page and adding routing to it depending on the payment type, or if the user has declared convictions. It does not cover other sorts of conviction checks yet.

* Scaffold renewal received form

* Set up page layout and content

* Remove unnecessary routes, tweak content

* Reroute users with declared convictions

If a user declares they have convictions, we need to do some extra checks before renewing their registration. So they should be routed to the renewal_received_form instead. In the future this will cover any registrations flagged by the entity matching service as well.

Because a user could end up on the page because of their payment method, having a pending conviction check, or both, we need to make sure they only see content relevant to them.

* Fix typo