Only display user's registrations in temp listing page (#151)
* Only display user's registrations in temp listing page

https://eaflood.atlassian.net/browse/WC-278

For testing purposes only, we have a page which lists all registrations in the database and lets users start the renewal process. We should limit this to just showing registrations belonging to the logged-in user account.

* Users can only read registrations belonging to them

Tweak users' abilities in CanCanCan so registrations are only visible if the account email matches. This stops other users' registrations from being listed in the temp page.

* Seed users from JSON

We want to be able to seed multiple users, so rather than cluttering up the seed script further, just add them all to a JSON file instead.

Currently we only read the email from JSON and then provide the same password based on an env var or a default.

* Clean up temporary registrations page

Minor copy changes to make this a bit clearer for testing, plus a nice big green 'RENEW' button.