Make sure can_renew checks are actually applied (#159)
* Make sure can_renew checks are actually applied

Our state machine for registrations has several checks to see if a registration can go through the renew action. However, we don't actually have this implemented anywhere (oops). This change makes sure that we are actually putting those checks into practice, so if a user bypasses the registration search in the frontend and accesses the renewal forms directly, they still can't renew something which should not be renewable.

* Check if reg can be renewed in FormsController

Whenever someone makes a GET or POST request to view or submit a form, we check to see if that registration can be renewed. If not, the user is redirected to an error.

* Capitalise registration states

For example, 'active' to 'ACTIVE'.

The status of a registration (in the metadata) was previously lowercase by default. However, when waste-carriers-services adds new registrations, it sets the state entirely in caps. So for our state checks to match, we have to capitalise all our state names.