Bug: Overseas users fail validation on 'check your details' (#171)
https://eaflood.atlassian.net/browse/WC-371

If you selected "Not in the United Kingdom" for your principal place of business, you got this error:

"Because your business type has changed, you cannot renew and must make a new registration instead"

Saying you're based overseas sets the business type to "overseas". However, it's valid for any existing type to change to this type, so this should be valid.