Pay for renewal using WorldPay (#167)
* Pay for renewal using WorldPay

https://eaflood.atlassian.net/browse/WC-117

Users should be able to pay for their renewal using WorldPay. We should also handle any payment failures or errors with WorldPay in a graceful way (eg without throwing an error or locking the user out of the process). Once a successful payment has been made using WorldPay, redirect to the 'renewal complete' or 'renewal application received' page (depending on whether a conviction check is pending).

* Build XML to send to Worldpay in correct format

Set up an XML builder using Nokogiri to build our request to Worldpay. Still need to fill in the data, but we know the structure is there.

Also add account_email to registration factory as we'll need this to build the XML for tests.

* Populate Worldpay XML with info from transient_registration

Fill in the missing info for the initial request sent to Worldpay (aside from the payment info and merchant id).

Fleshing out the address factories for testing also revealed a bug in prepopulating manual address forms - this commit also fixes that.

* Move XML-building into separate service

There were a lot of private methods just being used to build this XML, so it felt like time to move it off to its own class.

* Add Worldpay env vars

This allows us to include the merchant code in the XML (although we stub it for easier testing).

* Build new orders and order items

The types of order_items we need to deal with for now are the basic renewal charge, the additional charge for changing the reg type and the charge for any copy cards. This commit sets up methods to build a new order with the appropriate nested items.

* Add method to update finance_details balance

* Include order amount in Worldpay XML

* Generate order_code and include it in XML

The current app generates the order_code based on the current time, so I've stuck with that.

This means we should now be generating a complete XML file, so we can start testing whether Worldpay will accept it.

* Tweak stubbing for Worldpay XML test

The cost for orders and orderitems relies on environment variables, so for consistent testing it's better to just stub the value instead.

* Send initial request to WorldPay and parse the response

Now that we have the XML, we can send it to WorldPay and get back a response. This contains the link we need for redirecting our users, as well as a reference ID.

* Add env vars to Travis config

We were missing a value for the copy card cost. Also it expects some config for WorldPay or the tests fail.

* Redirect users to and from WorldPay

Use the URL we get from WorldPay to redirect users to the payment screens. Then, at the end, send them back to the app (either to an end screen or back to payment summary).

This change also renames financeDetails to finance_details as the irregular naming was causing issues. We will probably need to do this with other nested objects in the registration/transient registation collection, but this can be done as a separate PR later.

We also aren't actually persisting financeDetails yet, so this will need to be taken care of to actually store order and payment info.

* Store amounts in pence instead of pounds

* Set up payment objects

When we prepare for a Worldpay payment, set up a new payment object to store this information in the database.

Also a bit of test refactoring for clarity's sake.

* Make sure finance_details persist when created

* Validate success response from WorldPay before proceeding

When the user is sent to the success path after completing a WorldPay payment, make sure the params are as expected.

Additional work is needed here to validate the MAC.

* Improve test coverage, light refactoring

* Persist payments, and update after WorldPay success

When we create a payment, persist it in the database. When WorldPay returns a valid success response, update the payment to record that it has been successful. We also now only factor in successful payments when calculating the balance.

* Save payment data if payment fails

But only if the failure request is valid.

Also restructure the WorldpayController a bit to try to tidy things up.

* Update balance after successful payment

* Validate MAC when we get a success or failure from WorldPay

We store a MAC on WorldPay and in our .env file. We can use this to verify that the requests we get when a user is redirected back from WorldPay are valid (rather than someone attempting to fake a payment by URL-hacking).

This code was mostly cribbed from the frontend app. However, for a similar implementation in a more up-to-date verson of WorldPay, see:
http://support.worldpay.com/support/kb/gg/corporate-gateway-guide/content/hostedintegration/securingpayments.htm

* Use I18n for payment error messages

* Save date info after successful or failed payment

* Save registration_reference and comment for payments

This is for consistency with the existing system.

* Copy worldpay status to order after payment

* Store merchant code against order

* Save date_entered on payments

Also move some of the tests for update_after_worldpay method to the Payment model tests instead of the Worldpay controller tests.

* Save mac_code to payment in database

* Set date_created and date_last_updated for orders

* Set updated_by_user for orders and payments

Since we haven't yet implemented assisted digital registrations, this will only ever be the account email for the transient registration. So that's what we're using for now. Will require some changes later down the road.

* Only persist successful payments

Previously we were saving all payment objects when we set up our initial requests to Worldpay. However, to match the current system, we should only save payments which have been sucessfully processed.

However, unsuccessful payments do still update the status of the order.

* Move logic from Worldpay controller to service

Our controller was getting pretty bulky, so this commit shifts a lot of the logic back into the service where it belongs. This has a knock-on effect on the tests as well.

* Move Worldpay URL generation to new service

WorldpayService is now getting pretty big, so this commit moves the 8 methods involved in generating the redirect URL to a separate service.

* Move Worldpay success and failure validations to new service

This change further streamlines the WorldpayService by moving all the validation for successful and failed payments into a separate service. This should hopefully make the WorldpayService much more readable so it's easier to get an overview of the behaviour, with the nitty-gritty stored elsewhere.

* Refresh VCR cassettes and fix tests

This was initially done as the port for the address finder service had changed, and to keep all the cassette expiry times in sync. However, refreshing the cassettes also uncovered a few issues with unnecessary stubbing, which this commit also fixes.

* Store charge values as pence by default

Then convert them into pounds when we need to display them to the user. This makes the app more consistent with the rest of the service, and also more flexible - if we needed to charge a fee that wasn't a round number, we could now do it.